pyproject.toml
[project]
name = "lightkube-models"
dynamic = ["version"]
description = "Models and Resources for lightkube module"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Giuseppe Tribulato", email = "gtsystem@gmail.com" }
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.urls]
Homepage = "https://github.com/gtsystem/lightkube-models"
Repository = "https://github.com/gtsystem/lightkube-models"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/lightkube/models/__init__.py"

[tool.hatch.build.targets.sdist]
packages = ["src/lightkube/models", "src/lightkube/resources"]
sources = ["src/"]
ignore-vcs = true

[tool.hatch.build.targets.wheel]
packages = ["src/lightkube/models", "src/lightkube/resources"]
sources = ["src/"]
ignore-vcs = true

[dependency-groups]
compile = [
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown-include>=0.8.1",
    "mkautodoc>=0.2.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-material-extensions>=1.3.1",
    "ruamel.yaml>=0.18.0",
    "typer>=0.15.1",
]
